Update LAB_0204_Implement_SQL_Server_Resource_Provider.md
The change resolves the issue: SqlAdapter installation failed #10

Do not set the environment and perform the authentication for AzureStackAdmin, steps 4 to 7 of task 1. The authentication process is performed by the DeploySQLProvider.ps1 script.

The Lab hoster VM image already have Az account context. We don't need to remove the account.